changed how instructions are dispatched
removed the old model where the execution loop would attempt to generically decode arguments. that way wasn't going to work because there isn't a generic model. for the most part the logic works fine. but the address / data bus stuff is sometimes very specific to an (opcode, address mode) pair. this commit wraps up the 6502 tests and instructions (except for "illegal" instructions for now). on to 65c02.